requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
httpx>=0.28.0
redis>=5.2.0
python-dotenv>=1.0.0
